[DEV] added new target <module>-doc for building a documentation
Doxygen handles the documentation.
If the specified module defines LOCAL_DOXYFILE then the corresponding
doxyfile is used, otherwise, a doxyfile is generated on the fly with
defaul values for parameters which will scan recursively the module's
directory for source files and will generate documentation for all the
symbols and for all the languages.

Change-Id: I5c8ce2e3a8ad83faa60eb2003eb373745dbf2e5b