[DEV] Fix parsing of bison version
On my config, the bison version "bison (GNU Bison) 2.7.12-4996" is
not parsed correctly because of the trailing part with a digit and
thus, the whole string is considered instead of the numeric part.

This causes weird logs when check-version gets called on that string
as parenthesis are not handled properly :
/bin/bash: -c: line 0: syntax error near unexpected token `('
/bin/bash: -c: line 0: `expr bison (GNU Bison) 2.7.12-4996 \>= 2.5'

Fixing the regex fixes the issue for me.

Change-Id: Ia2dc178091b74f0bcc9e521e053668088f76043d
Reviewed-on: https://gerrit.parrot.biz/13623
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>
(cherry picked from commit 20623e46ceec61002fab35d526e2f87f4deb922b)