[DEV] Use a specific target for post-install action
It fixes an issue for users which want to have a POST_INSTALL rules
using installed files without knowing resulting filenames, for instance
with recipes like:
$(eval libs := $(wildcard $(TARGET_OUT_STAGING)/usr/lib/mylib_*))

Without a specific target for post-install action, definition of
LOCAL_xxx_CMD_POST_INSTALL written in local atom.mk and 'libs' variable will
be expanded when 'install' target is fired.
But at his stage, mylib_* files are not yet in staging since this
work is done in 'install' recipe resulting in the libs variable to be
empty.

So this commit fix it using a different target for post-install action
in order to depends of the result of the install one.

Change-Id: I4805b473a0c58a5d378f02bfd233659bc162d982
Reviewed-on: https://gerrit.parrot.biz/13906
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>