[DEV] Make custom macros preserve colons in argument list
exec-custom-macros now only substitutes the first colon in a macro
registration declaration to split the macro name from the arguments.

As a result, colons in the argument list are preserved as is.

The motivating use case for this change is being able to use a cpp
namespace (a::b::c) as an argument for a code generating macro

Change-Id: Id48cf3d7b197a17f9291ac9d72df5de92e757763
Reviewed-on: https://gerrit.parrot.biz/32060
Reviewed-by: Julien Thomas <julien.thomas@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>
(cherry picked from commit fc2708de83dc0d939c4a8fb447301f7229e38c61)