[DEBUG] add -march=armv7-a in ld flags to handle android gcc 'multilib' feature
The GCC provided through the android standalone toolchain for ARM is using the
'multilib' gcc feature (https://gcc.gnu.org/install/configure.html), which only
works correctly when both the compiler flags AND the linker flags specify the
architecture.

Here is the output of 'arm-linux-androideabi-gcc -print-multi-lib' for the
toolchain used to compile FF3 on armeabi_v7a variant:
    .;
    armv7-a;@march=armv7-a
    thumb;@mthumb
    armv7-a/thumb;@march=armv7-a@mthumb
It shows that there are 4 variants available and where to find their respective
standard libraries.

Not passing the -march flag at link stage results in linking the wrong library
(the default lib which correspond to armv5 abi) with the binary or shared
object. This is especially the case with binary that are to be linked against a
non standard libstdc++ (STL port) as the link is static.

This patch only addresses the armv7-a  as to my knowledge only ARM architecture
support 2 different abi (see http://developer.android.com/ndk/guides/abis.html),
but this may change in the future so maybe this patch has to be reworked.

Change-Id: I39305bdeb6df58ba0442498932a031785349d022
Reviewed-on: https://gerrit.parrot.biz/34787
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>