[DEV] Make possible to build kernel modules for different kernel release
This is useful to build a kernel module within a docker container.
'uname -r' will return the release of the kernel of the host machine and
build will fail if another version of linux headers is installed in the
docker image.

It's not be possible to install the package of the linux headers
of the host machine if the the docker image is based on a different
release of linux.

Change-Id: I49c76d84504c5dbffa069033c6bf9cab0f179952
Reviewed-on: https://gerrit.parrot.biz/37223
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>
(cherry picked from commit 66cc341c5072ac2240ff32865ab80cdbf01bd87d)