[DEV] Support using alchemake as part of parallel build
If alchemake is invoked from another makefile, and that makefile is
build in parallel (-j flag), then the parent make process will open a
pipe to let the child make process connect to jobserver.

However, the pipe was closed when alchemake.py spawned the child make
process. To make it work, we need to keep existing file descriptors open
when creating the child process.

Change-Id: I332eb6f49f22b1485c32a2f3adf84e318b2ffc3c
Reviewed-on: https://gerrit.parrot.biz/54862
Reviewed-by: Guillaume Blanc <guillaume.blanc@parrot.com>
Reviewed-by: Aymen Bouaziz <aymen.bouaziz@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>