[DEV] add support for zip compressed image format
This format is needed by the sphinx team because the zip linux command
supports adding comment for the entire archive. This field is used to
add the UUID for ext images in the .ZIP file comment field of the
"end of central directory record" section (end of file). For other image
formats (tar and cpio), the comment field remains empty.

Change-Id: I8b7b1754bc61aaa077b6f31a24c3950825607eb6
Reviewed-on: https://gerrit-doc.parrot.biz/80943
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>