[DEV] Split the generated Android.mk in SDK
Create a new Android-static.mk, which only contains the static
libraries. Included by Android.mk for compatibility

When an included Android.mk declares a "PREBUILT_SHARED_LIBRARY", it
will be copied into the ndk-build out folder (and then into the APK
file), even if it is not actually used by the app. To avoid this, apps
which only use static libraries can use the Android-static.mk file.

Change-Id: I724818af5c8a9333a580eab80e75ee782490fc2a
Reviewed-on: https://gerrit-doc.parrot.biz/84947
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>