[DEV] fix ldconfig assert (with new libc locales)
libc v2.27 add new locale elements causing assert issue during the 'final'
step of the build process (but the issue is not 2.27 specific).

The issue is that alchemy-packaged ldconfig is a frozen old version
(because of cross-compiling issues: see ldconfig/README) and is not in
sync with the host locale database (/usr/lib/locale/locale-archive).

When new locale elements are added to the host database (as it is the
case with libc v2.27), hardcoded array size in the old ldconfig is not
big enough to store these elements and an assert is catched
(always better than an array overflow).

This patch set locale of all categories to "C" (with LC_ALL=C) before
executing ldconfig. Because when a category locale is "C" ldconfig do not
read locale elements from host database but use its hardcoded "C" locale
elements avoiding the "not in sync" issue.

Change-Id: I4d77902e925fbf5247eb1b43fdd9a7622ffb03b5
Reviewed-on: https://gerrit-doc.parrot.biz/87693
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>