[DEVMINOR] Prevent capturing output from cd
When using a non-null CDPATH environment variable, "cd" echoes the
directory being switched to. This string is captured by the subshell
command, which results in a PATH variable containing an invalid entry:

    "<output-of-cd> <output-of-pwd>:$PATH"

instead of:

    "<output-of-pwd>:$PATH"

Redirecting the output of cd to /dev/null fixes that.

Alternatively, use realpath (or readlink) if that is possible:

    $(dirname $(realpath -P -- ...))

Change-Id: I4aaa12b991c9f6c40ae91707bfd58cc39c61da61
Reviewed-on: https://gerrit-doc.parrot.biz/92913
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>