[DEBUG] Fix compatibility with pkgconf
By default, pkg-config looks for .pc files in its primary directory on
the host. This directory can be overridden with the environment variable
PKG_CONFIG_LIBDIR.

With pkg-config, setting this variable to the empty string is enough to
remove the default libdir from the list of primary directories.
However, with pkgconf, another implementation of pkg-config, doing so
does not have any effect (tested with pkgconf 1.5.4).

This patch sets the PKG_CONFIG_LIBDIR environment variable to the
staging dirextory, where we can reasonably expect that no .pc files from
the host will be present.

Change-Id: I8d20c5a6d2d74d769494f7cd5df557d4b2b37d39
Reviewed-on: https://gerrit-doc.parrot.biz/101264
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>