[DEV] avoid duplicate kernel module [2/2]
The commit that manages install/remove directory
for kernel module implies a duplication of
this kernel module.

The kernel module was copied from /lib/modules/4.9.37/extra/
to the target directory.

This new commit creates a symbolic link instead off copying it
without breaking the compatibility with depmod tool
that manages kernel module dependencies.

Change-Id: I7af7e7d07449b78429ef620922a0492cf4c86b67
Reviewed-on: https://gerrit-doc.parrot.biz/103629
Reviewed-by: Benjamin Bagland <benjamin.bagland@parrot.com>
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>