[DEV] avoid duplicate kernel module [1/2]
This bug was introduced by the previous commit that
manages install/remove directory for kernel module.

This features relies on the definition of the variable
LOCAL_DESDIR later assigned to PRIVATE_DESDIR.

When LOCAL_DESDIR is not defined, PRIVATE_DESDIR is empty
and as a result the following test in bash without quote
around $(PRIVATE_DESTDIR) is true.

Test in bash:
	if [ -n $(PRIVATE_DESTDIR) ]; then
	...

This leads to the same result wether LOCAL_DESDIR is defined
or not, to know duplicate the kernel module.

Change-Id: Ie15d18f0ea4cf726adc17f640b837599f29ca402
Reviewed-on: https://gerrit-doc.parrot.biz/103758
Reviewed-by: Benjamin Bagland <benjamin.bagland@parrot.com>
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>