[DEV] Fix checkpatch.pl to detect more lines > 80 columns
Removed tests:
-> If the previous line contained "/**", the current line was assumed
to be a kerneldoc line, which (a long time ago) had to be on a single
line. Since this restriction no longer makes sense, we can remove
this.
-> If the line is a long documentation line of an indentifier (" *
@something something..."). We can also remove it as we limit all
comments lines to 80 columns.
--> If the line match a kernel log function (+ some perl regex magic),
the line is allowed to go past 80 columns. This is not useful in our
use case since we don't use this patch for checking kernel code.

Added test:
-> If the line contains an URL (e.g. "https://some_long_url.com"),
then the line is allowed to be longer than 80 columns, this avoid
breaking the URLs in comments & defines.

Change-Id: Iab64ab89dbdac19172dec97cafe94690ef378872
Reviewed-on: https://gerrit-doc.parrot.biz/104287
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>