[DEV] Reduce amount of repeated code to locate gdbserver
Instead of stacking calls to $(wildcard), it's possible
to feed all the paths to $(wildcard) and keep only the
first result (if there's more than one matching gdbserver,
which is very unlikely).

And, by constructing the paths to lookup with $(addprefix)
and $(addsuffix), it's possible to reduce more the amount
of repeated text.

Change-Id: Id928cb093e781f615bc6b2ecfa0308e8bb856017
Reviewed-on: https://gerrit-doc.parrot.biz/118334
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>