[DEBUG] Fix OSError from possibly empty directory creation
This patch fixes an issue that has been exposed with commit
49708d802ebc5209cff74bb6dff7f60ce1f9fef7, which switch the module config
path from absolute to relative path.

In some projects the module config path is located at the root of the
project, hence os.path.dirname(configFilePath) returns an empty
string used to create empty directory leading to an OSError.

Change-Id: I5bf7f805cc6f08df2b8eef5ae4c478972b4f3020
Reviewed-on: https://gerrit-doc.parrot.biz/120445
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Nicolas Fugier <nicolas.fugier@parrot.com>
Reviewed-by: Hugo Sarbia <hugo.sarbia.ext@parrot.com>
Reviewed-by: Pierre-Moana Levesque <pierre-moana.levesque@parrot.com>