[DEV] codeformat: inherit as much as possible of the original file
Using mktemp then moving the temporary file to replace
the original source file has some drawbacks.

First, if the source code before formatting and after
formatting is the same, replacing the file with a new
one with the same content is pointless, triggering
refresh(es) in various editors/IDEs, and unneeded
rebuilds, as source file is now younger than its object
file.

Here, if the content is the same, the temporary file
won't be used: it won't replace the original file anymore.
In the other hand, while being semantically identical,
if the content is different, no effort is made to prevent
rebuilding the file: modification time from original file
is not retained, allowing the compiler to check the result
of the formatting.

Second, temporary files are created with very limited
permissions (0600), which differ from the original files
(usually 0644), breaking some specific use cases.

As POSIX doesn't specify a mean to retrieve a representation
of the permission that can be fed back to chmod, it's
rather complicated to copy the permission from one file
to another.

Except with GNU coreutils's chmod, which has the very
nice --reference option.

Unfortunately this version of tool is not available on
BSD based systems such as macOS, so other methods must
be used.

Here, multiple methods are attempted, and hopefully,
one will work.

Change-Id: I0b28454cc2597053da18ceb84cf19ef103af748d
Reviewed-on: https://gerrit-doc.parrot.biz/120924
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yann Droneaud <yann.droneaud.ext@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>