[DEV] Search libasan in <sysroot>/usr/lib too
For USE_ADDRESS_SANITIZER to work, libasan library is required
to be available in staging to be found at link time, and in root
file system at run time.

As it's quite usual to have the library in /usr/lib{,64}, search
it there, allowing address sanitizer to be used with
parrot-tools-gcc-aarch64-himix100-linux toolchain.

Change-Id: I20922592f06db64323b0e5d841b26f0c285144a0
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/126841
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yann Droneaud <yann.droneaud.ext@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>