[DEV] clang: llvm-link is not a valid linker
llvm-link is meant to link LLVM bytecode, thus it cannot
be used to link ELF object files into a shared library.

At least since LLVM 4.0, the project provides ld.lld as
a drop-in replacement to ld, and it aims to be compatible
with GNU ld.

So ld.lld should be used instead of llvm-link.

Beware, on most Linux distribution, such as Debian, Fedora,
Ubuntu, etc. LLVM's ld must be installed as it's not
provided by clang or llvm packages. Install it with
command such as:

- Debian/Ubuntu:  apt install lld
- Fedora:         dnf install lld

Change-Id: I069164a991ebaea7437823d14071830ebfc4680c
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/154407
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>