[DEV] LINUX_MODULE: remove dead link before installing module
When rebuilding and installing a module already built and installed
for a different Linux kernel version, a failure is reported:

      INSTALL .../out/<product>/build/<module>/obj/<module>.ko
    ln: failed to create symbolic link '.../out/<product>/staging/lib/.../<module>.ko': File exists
    make: *** [.../build/alchemy/classes/GENERIC/rules.mk:129: .../out/<product>/build/<module>/<module>.installed.stamp] Error 1

Link exists but it's dead as it points to the previous kernel
version's modules directory:

    $ file .../out/<product>/staging/lib/.../<module>.ko
    .../out/<product>/staging/lib/.../<module>.ko: broken symbolic link to ../../modules/4.9.y/extra/<module>.ko

The error happens because shell's test -e does follow the symbolic
link and return false, skipping the previous symlink's removal.

As rm -f won't report an error if the target file doesn't exist,
it can be done unconditionally, removing the need for testing
for the presence of the symbolic link.

Change-Id: I3eb97c0df121d3238ffd63b15e89d65ec2ca1f7a
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/158485
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>