[DEV]linux/native: change "libjpeg-turbo" prebuilt to match our module
Note about names:
- `libjpeg-turbo' is the name of one of the most common libjpeg
implementation.
- `libturbojpeg' is the name of another jpeg library (with a
different API). This API is also provided by libjpeg-turbo, but not
by default.

This means that the name "libjpeg-turbo" for a package name is not
really clear, but changing this would break a lot of packages, so it
seems better to live with it.

Since the libjpeg-turbo non-prebuilt package actually builds both
implementations by default, and exports both `-ljpeg' and
`-lturobjpeg' linker flags, this patch changes the prebuilt module to
have the same behavior: it is only available when both libraries are
available, and export both libjpeg and libturbojpeg c/ld flags.

Change-Id: Ief7591967f063af1bc5a1904eb002971a79038a0
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/180369
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Mathieu Le Mauff <mathieu.lemauff@parrot.com>
Reviewed-by: Christian Rosalie <christian.rosalie@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>