[DEBUG] Fix LOCAL_REQUIRED_MODULES
Alchemy would never include runtime dependencies when genereting the
input arguments for confwrapper.py.
This was caused by the check for __CONFIG_DISABLE_RUNTIME_DEPS, which
was implemented as follow:
$(if $(__CONFIG_DISABLE_RUNTIME_DEPS),value_if_true,value_if_false)

Make's $(if COND,THEN,ELSE) function, evaluates as THEN is COND is a
non-empty string, otherwise it evaluates as ELSE (or an empty string if
ELSE is omitted).

Since __CONFIG_DISABLE_RUNTIME_DEPS is assigned either "0" or "1", the
condition was always true, so the runtime dependencies were always
ignored.
Because of this, the LOCAL_REQUIRED_MODULES variables had no effect.

This commit filters __CONFIG_DISABLE_RUNTIME_DEPS so that COND evaluates
to "1" or the empty string.

Change-Id: I492ed60bad0e3c684da72e1e371b747499ebe1fb
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/190552
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>