[DEV] Improve USE_xxx_SANITIZER syntax
Current syntax:
USE_xxx_SANITIZER can either contain "1" (enabled on all module), or
a list of modules on which the sanitizer is enabled

This patch proposes the following syntax (space-separated):
- `1' still enables everything, in order to keep compatibility
- `module' enabled the sanitizer for the given module (as per current
syntax)
- `-module' forbids the sanitizer for the given module
- `module*' enables the sanitizer for the given module, and all of
its (direct & indidect) dependencies
- `-module*' forbids the sanitizer for the given module, and all of
its (direct & indidect) dependencies

This allows for more complex sequences:

`1 -moduleA -moduleB' enables the sanitizer for everything except
moduleA and moduleB. This can be useful to disable the sanitizers on
external libraries, or on libraries which have known issues

`moduleA*' enables the sanitizer for moduleA and all of its
dependencies. This can be useful to test the sanitizer on a given
binary without having to rebuild everything

`moduleA* -moduleB*' enables the sanitize for moduleA and all of its
dependencies, as long as they're not dependencies of moduleB (nor
moduleB itself). This can be useful on target to enable the sanitizer
on a given binary, as deep as possible, without touching any library
used by anoter binary.

Change-Id: I4dc2660e13fdc6a1f749eec083230ef8e3407f28
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/191887
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>