[DEV] Allow to customize veritysetup format params
Add the TARGET_IMAGE_VERITY_FORMAT_PARAMS variable, which allows to
customize the parameters passed to the `veritysetup format` command.

A new dm-verity-config.txt configuration file will be generated in
addition to the U-boot script. It should be easier to parse and convert
to another format.

Also, since this change, we no longer pass default parameters
(--data-block-size=1024) to veritysetup, so you may need to adjust your
product.mk if you were using dm-verity for your rootfs image.

Also add the TARGET_IMAGE_VERITY_OPTIONS to the list in variables.mk, as
it was missing.

Change-Id: If92caf0932663d72ddae0a51e7d3e5d6a69b7605
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/226987
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Benjamin Bagland <benjamin.bagland@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>