[DEV] Header path normalization rework
The choices between -isystem and -I should be made, for instance,
in toolchains/.
If unspecified, -I is assumed.

The version/flavour-checking logic in the normalization function
definitions is IMHO quite odd, so I haven't kept it.

Make sure to avoid using spaces when using -I or -isystem., otherwise
this will confuse the normalization functions.

Change-Id: Ie81af6724b5cdaba3bbba9155037451fa7ed6b86
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/246422
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Benjamin Bagland <benjamin.bagland@parrot.com>