[DEV] Add missing nconfig
Linux-headers is a global prerequisite, so Alchemy always reads the file
where its rules are defined.
Before linux-headers was split to rules-linux-headers.mk, that was
rules-linux.mk. Coincidently, that's where linux-nconfig is defined,
which makes -A linux-nconfig accidentally work.

After the split, rules-linux-headers.mk was read instead and the
linux-nconfig rule in rules-linux.mk was never read.

Change-Id: I5c74bfb917f63780f9da060ee699cda2e9197813
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/250533
Reviewed-by: Matthias Rabault <matthias.rabault@parrot.com>