[DEV] Remove linux-% rule
This rule matches [...]/linux-headers.revision, which will call
the Linux Makefile with [...]/headers.revision.
This is harmless but it causes spurious logs. After the linux-headers
split, rules-linux.mk isn't necessarily read anymore, which prevents
this rule from working if -A linux-foo is run, for instance.

Since some people use linux-dtbs, this commit adds a rule for linux-dtbs
and tweaks is-module-in-make-goals to make Alchemy read linux-rules.mk
when linux-dtbs is passed to Alchemy.

Change-Id: Ie477c1f5fa4b23018d6745e095e035453b5afbec
Reviewed-on: https://gerrit-doc.parrot.biz/c/build/Alchemy/+/250568
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Benjamin Bagland <benjamin.bagland@parrot.com>